backend/requirements.txt
django
django-cors-headers
djangorestframework
bs4
requests